# openapi.yaml
openapi: 3.1.0
info:
  title: BuyWhere Catalog API
  description: |
    Agent-native product catalog API for AI agent commerce.
    Query millions of products across Southeast Asia.
  version: 1.0.0
  contact:
    name: BuyWhere Engineering
    url: https://buywhere.ai
  license:
    name: Proprietary
servers:
  - url: https://api.buywhere.ai
    description: Production
  - url: https://staging-api.buywhere.ai
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /v1/ingestion/stats:
    get:
      tags:
        - ingestion
      summary: Get ingestion statistics
      description: |
        Aggregate statistics across all ingestion runs.
        Includes total runs, completion rates, and row counts.
        Cached for 60 seconds.
        Rate limit: 100 requests/minute per API key.
      operationId: getIngestionStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Ingestion statistics
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 99
            X-RateLimit-Reset:
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionStats'
              example:
                total_runs: 1250
                completed_runs: 1180
                failed_runs: 45
                running_runs: 25
                total_rows_inserted: 4500000
                total_rows_updated: 1200000
                total_rows_failed: 8500
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/status:
    get:
      tags:
        - status
      summary: Get catalog health and status
      description: |
        Returns catalog health information including total product count,
        per-platform counts, last update times, and progress toward goals.
        Rate limit: 100 requests/minute per API key.
      operationId: getCatalogStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Catalog status
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 99
            X-RateLimit-Reset:
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogStatus'
              example:
                total_unique_products: 452300
                goal_million: 1
                progress_percent: 45.23
                platforms:
                  - source: shopee_sg
                    product_count: 150000
                    last_updated: '2026-04-03T10:30:00Z'
                  - source: lazada_sg
                    product_count: 120000
                    last_updated: '2026-04-03T09:15:00Z'
                ingestion_runs:
                  - source: shopee_sg
                    last_run_at: '2026-04-03T10:00:00Z'
                    last_run_status: completed
                    last_rows_inserted: 950
                    last_rows_updated: 48
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/status/scrapers:
    get:
      tags:
        - status
      summary: Get scraper health status per platform
      description: |
        Returns health status for each platform's scraping pipeline.
        Includes last run time, row counts, and error information.
        Rate limit: 100 requests/minute per API key.
      operationId: getScraperHealth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Scraper health report
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 99
            X-RateLimit-Reset:
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScraperHealthReport'
              example:
                generated_at: '2026-04-03T10:30:00Z'
                total_scrapers: 6
                healthy_count: 5
                unhealthy_count: 1
                scrapers:
                  - source: shopee_sg
                    last_run_at: '2026-04-03T10:00:00Z'
                    last_run_status: completed
                    last_rows_inserted: 950
                    last_rows_updated: 48
                    last_rows_failed: 2
                    product_count: 150000
                    is_healthy: true
                    hours_since_last_run: 0.5
                    error_message: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/catalog/stats:
    get:
      tags:
        - catalog
      summary: Get catalog statistics
      description: |
        Returns catalog statistics including total products by platform and category.
        Cached for 5 minutes.
        Rate limit: 100 requests/minute per API key.
      operationId: getCatalogStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Catalog statistics
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 99
            X-RateLimit-Reset:
              schema:
                type: integer
                example: 1709568060
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogStatsReport'
              example:
                generated_at: '2026-04-03T10:30:00Z'
                total_products: 452300
                by_platform:
                  shopee_sg: 150000
                  lazada_sg: 120000
                  carousell: 80000
                  qoo10: 50000
                by_category:
                  Electronics: 180000
                  Fashion: 90000
                  Home & Living: 60000
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API key
      description: |
        Authentication via API key or JWT token.
        - For direct API keys: use the key prefixed with `bw_` as-is
        - For JWT tokens: obtain via OAuth flow, contains `key_id` claim
        All API requests must include this header.
  schemas:
...
    IngestionStats:
      type: object
      required:
        - total_runs
        - completed_runs
        - failed_runs
        - running_runs
        - total_rows_inserted
        - total_rows_updated
        - total_rows_failed
      properties:
        total_runs:
          type: integer
          description: Total number of ingestion runs
          example: 1250
        completed_runs:
          type: integer
          description: Number of successfully completed runs
          example: 1180
        failed_runs:
          type: integer
          description: Number of failed runs
          example: 45
        running_runs:
          type: integer
          description: Number of currently running runs
          example: 25
        total_rows_inserted:
          type: integer
          description: Total rows inserted across all runs
          example: 4500000
        total_rows_updated:
          type: integer
          description: Total rows updated across all runs
          example: 1200000
        total_rows_failed:
          type: integer
          description: Total rows failed across all runs
          example: 8500
    PlatformProductCount:
      type: object
      required:
        - source
        - product_count
      properties:
        source:
          type: string
          description: Source platform
          example: shopee_sg
        product_count:
          type: integer
          description: Number of active products
          example: 150000
        last_updated:
          type: string
          format: date-time
          nullable: true
          description: Last product update timestamp
          example: '2026-04-03T10:30:00Z'
    IngestionRunInfo:
      type: object
      properties:
        source:
          type: string
          nullable: true
          description: Source platform
          example: shopee_sg
        last_run_at:
          type: string
          format: date-time
          nullable: true
          description: Last ingestion run timestamp
          example: '2026-04-03T10:00:00Z'
        last_run_status:
          type: string
          nullable: true
          description: Last run status
          example: completed
        last_rows_inserted:
          type: integer
          nullable: true
          description: Rows inserted in last run
          example: 950
        last_rows_updated:
          type: integer
          nullable: true
          description: Rows updated in last run
          example: 48
    CatalogStatus:
      type: object
      required:
        - total_unique_products
        - platforms
        - ingestion_runs
        - goal_million
        - progress_percent
      properties:
        total_unique_products:
          type: integer
          description: Total number of unique active products
          example: 452300
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/PlatformProductCount'
        ingestion_runs:
          type: array
          items:
            $ref: '#/components/schemas/IngestionRunInfo'
        goal_million:
          type: number
          format: float
          description: Goal in millions of products
          example: 1
        progress_percent:
          type: number
          format: float
          description: Progress toward goal percentage
          example: 45.23
...
    ScraperHealth:
      type: object
      required:
        - source
        - product_count
        - is_healthy
      properties:
        source:
          type: string
          example: shopee_sg
        last_run_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-04-03T10:00:00Z'
        last_run_status:
          type: string
          nullable: true
          example: completed
        last_rows_inserted:
          type: integer
          nullable: true
          example: 950
        last_rows_updated:
          type: integer
          nullable: true
          example: 48
        last_rows_failed:
          type: integer
          nullable: true
          example: 2
        product_count:
          type: integer
          example: 150000
        is_healthy:
          type: boolean
          example: true
        hours_since_last_run:
          type: number
          format: float
          nullable: true
          example: 0.5
        error_message:
          type: string
          nullable: true
          example: null
    ScraperHealthReport:
      type: object
      required:
        - generated_at
        - scrapers
        - total_scrapers
        - healthy_count
        - unhealthy_count
      properties:
        generated_at:
          type: string
          format: date-time
          example: '2026-04-03T10:30:00Z'
        scrapers:
          type: array
          items:
            $ref: '#/components/schemas/ScraperHealth'
        total_scrapers:
          type: integer
          example: 6
        healthy_count:
          type: integer
          example: 5
        unhealthy_count:
          type: integer
          example: 1
    CatalogStatsReport:
      type: object
      required:
        - generated_at
        - total_products
        - by_platform
        - by_category
      properties:
        generated_at:
          type: string
          format: date-time
          example: '2026-04-03T10:30:00Z'
        total_products:
          type: integer
          example: 452300
        by_platform:
          type: object
          additionalProperties:
            type: integer
          example:
            shopee_sg: 150000
            lazada_sg: 120000
            carousell: 80000
            qoo10: 50000
        by_category:
          type: object
          additionalProperties:
            type: integer
          example:
            Electronics: 180000
            Fashion: 90000
            Home & Living: 60000
...
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Invalid or revoked API key
          example:
            detail: Invalid or revoked API key
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
          example: 60
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: 'Rate limit exceeded: 1000/minute'
          example:
            detail: 'Rate limit exceeded: 1000/minute'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Internal server error
          example:
            detail: Internal server error